requirements.txt
annotated-types==0.7.0
anyio==4.9.0
atproto==0.0.61
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
cryptography==44.0.3
dnspython==2.7.0
Flask==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ImageHash==4.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
libipld==3.0.1
MarkupSafe==3.0.2
numpy==2.2.5
pillow==11.2.1
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
PyWavelets==1.8.0
requests==2.32.3
scipy==1.15.3
sniffio==1.3.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
websockets==13.1
Werkzeug==3.1.3
